remove diff_test.go
