+++ raft/util.go
@@ -15,7 +15,13 @@
 package raft
 
 import (
+	"fmt"
+	"io"
+	"io/ioutil"
+	"os"
+	"os/exec"
 	"sort"
+	"strings"
 
 	pb "github.com/pingcap-incubator/tinykv/proto/pkg/eraftpb"
 )
@@ -68,6 +74,47 @@ func nodes(r *Raft) []uint64 {
 	}
 	sort.Sort(uint64Slice(nodes))
 	return nodes
+}
+
+func diffu(a, b string) string {
+	if a == b {
+		return ""
+	}
+	aname, bname := mustTemp("base", a), mustTemp("other", b)
+	defer os.Remove(aname)
+	defer os.Remove(bname)
+	cmd := exec.Command("diff", "-u", aname, bname)
+	buf, err := cmd.CombinedOutput()
+	if err != nil {
+		if _, ok := err.(*exec.ExitError); ok {
+			// do nothing
+			return string(buf)
+		}
+		panic(err)
+	}
+	return string(buf)
+}
+
+func mustTemp(pre, body string) string {
+	f, err := ioutil.TempFile("", pre)
+	if err != nil {
+		panic(err)
+	}
+	_, err = io.Copy(f, strings.NewReader(body))
+	if err != nil {
+		panic(err)
+	}
+	f.Close()
+	return f.Name()
+}
+
+func ltoa(l *RaftLog) string {
+	s := fmt.Sprintf("committed: %d\n", l.committed)
+	s += fmt.Sprintf("applied:  %d\n", l.applied)
+	for i, e := range l.allEntries() {
+		s += fmt.Sprintf("#%d: %+v\n", i, e)
+	}
+	return s
 }
 
 // TODO: Delete Start
